requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# TTS Libraries
gtts==2.4.0

# AI/ML Libraries - OpenAI
openai>=1.3.0
# Voice Cloning Dependencies (SV2TTS)
torch>=2.0.0
torchaudio>=2.0.0
librosa==0.10.1
soundfile==0.12.1
numpy>=1.24.3
scipy
matplotlib
umap-learn
scikit-learn
webrtcvad
inflect
unidecode

# Audio processing
pydub==0.25.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.0
Pillow==10.1.0
requests==2.31.0
tqdm

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Configuration
PyYAML==6.0.1
click==8.1.7